Added hash check to see whether solute changes or not
--HG--
branch : refactor

+++ hash_real.c
@@ -2,11 +2,11 @@
 #include <stddef.h>
 #include <string.h>
 
-void hash_double_(const double *v_, const size_t *elms_, uint64_t *o_)
+void hash_double_(const double *v_, const int *elms_, uint64_t *o_)
 {
   uint64_t x = 0, buf;
   size_t i, elms;
-  elms = *elms_;
+  elms = (size_t)*elms_;
   for(i = 0; i < elms; ++i){
     memcpy(&buf, &v_[i], sizeof(double));
     x = ((x << 7) | (x >> 57)) ^ buf;
@@ -14,12 +14,12 @@ void hash_double_(const double *v_, const size_t *elms_, uint64_t *o_)
   *o_ = x;
 }
 
-void hash_float_(const float *v_, const size_t *elms_, uint64_t *o_)
+void hash_float_(const float *v_, const int *elms_, uint64_t *o_)
 {
   uint64_t x = 0;
   uint32_t buf;
   size_t i, elms;
-  elms = *elms_;
+  elms = (size_t)*elms_;
   for(i = 0; i < elms; ++i){
     memcpy(&buf, &v_[i], sizeof(float));
     x = ((x << 7) | (x >> 57)) ^ (uint64_t) buf;
